Abstracted out formats in src/setup.py
Cleanups:
- Renamed classifier/ to src/
- Relevant changes to setup.py
