+++ src/config.py
@@ -1,0 +1,14 @@
+# Formats should be added below with respective directory name.
+formats = {
+    'Music': ['.mp3', '.aac', '.flac', '.ogg', '.wma', '.m4a', '.aiff'],
+    'Videos': ['.flv', '.ogv', '.avi', '.mp4', '.mpg', '.mpeg', '.3gp',
+               '.mkv', '.ts'],
+    'Pictures': ['.png', '.jpeg', '.gif', '.jpg', '.bmp', '.svg', 'webp',
+                 '.psd'],
+    'Archives': ['.rar', '.zip', '.7z', '.gz', '.bz2', '.tar', 'dmg', '.tgz'],
+    'Documents': ['.txt', '.pdf', '.doc', '.docx', '.xls', '.xlsv', '.xlsx',
+                  '.ppt', '.pptx', '.ppsx', '.odp', '.odt', '.ods', '.md',
+                  ',json', '.csv'],
+    'Books': ['.mobi', '.epub'],
+    'RPMPackages': ['.rpm']
+}
